estrutura do código atualizada
