--- src/GUI/Terminal.java
@@ -1,9 +1,0 @@
-
-public class Terminal {
-
-	public static void main(String[] args) {
-		
-
-	}
-
-}
