--- src/models/TipoDespesa.java
@@ -1,7 +1,0 @@
-package models;
-
-public enum TipoDespesa {
-    Escola,
-    Viagem,
-    Outro
-}
