--- src/test/GerenciadorTest.java
@@ -1,41 +1,0 @@
-import junit.framework.TestCase;
-
-public class GerenciadorTest extends TestCase {
-	
-	@Test
-	public void testInsereUsuario() {
-		
-		
-	}
-	
-	@Test
-	public void testRemoveUsuario() {
-		
-	}
-	
-	@Test
-	public void testInsereSaldo() {
-		
-	}
-	
-	@Test
-	public void testAtualizaSaldo() {
-		
-	}
-	
-	@Test
-	public void testInsereDespesa() {
-		
-	}
-	
-	@Test
-	public void testRemoveDespesa() {
-		
-	}
-	
-	@Test
-	public void testMostraExtrato() {
-		
-	}
-	
-}
